+++ deepfence_utils/postgresql/postgresql-db/models.go
@@ -18,7 +18,7 @@ type ApiToken struct {
 	CompanyID       int32
 	GroupID         int32
 	RoleID          int32
-	CreatedByUserID int32
+	CreatedByUserID int64
 	CreatedAt       time.Time
 	UpdatedAt       time.Time
 }
